configs: rename configs.ecu_info/proxy_info modules to _ecu_info, _proxy_info. while expose ecu_info, proxy_info objects on configs module
