+++ src/otaclient/boot_control/configs.py
@@ -18,7 +18,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 
 from otaclient.app.configs import BaseConfig
-from otaclient.configs.ecu_info import BootloaderType
+from otaclient.configs import BootloaderType
 
 
 @dataclass
